Refactor: libpacemaker: move update_colo_start_chain() to pcmk_sched_colocation.c
... and rename to pcmk__block_colocated_starts(). Also move a static function
used only by it.
